### 2023-07-22 #5675 @editorialbot
[PRE REVIEW]: Using the tidyverse with terra objects: the tidyterra package
**Submitting author:** @dieghernan (<a href="http://orcid.org/0000-0001-8457-4658">Diego Hernangómez</a>)
**Repository:** https://github.com/dieghernan/tidyterra
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.4.0
