+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   faqs: faqs.html
   palettes: palettes.html
   tidyterra: tidyterra.html
-last_built: 2023-07-18T22:46Z
+last_built: 2023-07-22T21:56Z
 urls:
   reference: https://dieghernan.github.io/tidyterra/reference
   article: https://dieghernan.github.io/tidyterra/articles
