+++ vignettes/tidyterra.Rmd
@@ -51,7 +51,20 @@ Load {tidyterra} with additional libraries of the {tidyverse}:
 
 ```r
 library(tidyterra)
+#> 
+#> Attaching package: 'tidyterra'
+#> The following object is masked from 'package:stats':
+#> 
+#>     filter
 library(dplyr)
+#> 
+#> Attaching package: 'dplyr'
+#> The following objects are masked from 'package:stats':
+#> 
+#>     filter, lag
+#> The following objects are masked from 'package:base':
+#> 
+#>     intersect, setdiff, setequal, union
 library(tidyr)
 ```
 
@@ -95,10 +108,17 @@ library(terra)
 f <- system.file("extdata/cyl_temp.tif", package = "tidyterra")
 
 temp <- rast(f)
-#> Error: [rast] filename is empty. Provide a valid filename
 
 temp
-#> Error in eval(expr, envir, enclos): object 'temp' not found
+#> class       : SpatRaster 
+#> dimensions  : 87, 118, 3  (nrow, ncol, nlyr)
+#> resolution  : 3881.255, 3881.255  (x, y)
+#> extent      : -612335.4, -154347.3, 4283018, 4620687  (xmin, xmax, ymin, ymax)
+#> coord. ref. : World_Robinson 
+#> source      : cyl_temp.tif 
+#> names       :   tavg_04,   tavg_05,  tavg_06 
+#> min values  :  1.885463,  5.817587, 10.46338 
+#> max values  : 13.283829, 16.740898, 21.11378
 
 mod <- temp %>%
   select(-1) %>%
@@ -106,10 +126,17 @@ mod <- temp %>%
   relocate(newcol, .before = 1) %>%
   replace_na(list(newcol = 3)) %>%
   rename(difference = newcol)
-#> Error in eval(expr, envir, enclos): object 'temp' not found
 
 mod
-#> Error in eval(expr, envir, enclos): object 'mod' not found
+#> class       : SpatRaster 
+#> dimensions  : 87, 118, 3  (nrow, ncol, nlyr)
+#> resolution  : 3881.255, 3881.255  (x, y)
+#> extent      : -612335.4, -154347.3, 4283018, 4620687  (xmin, xmax, ymin, ymax)
+#> coord. ref. : World_Robinson 
+#> source(s)   : memory
+#> names       : difference,   tavg_05,  tavg_06 
+#> min values  :   2.817647,  5.817587, 10.46338 
+#> max values  :   5.307511, 16.740898, 21.11378
 ```
 
 On the previous example, we had:
@@ -192,8 +219,12 @@ ggplot() +
     palette = "muted",
     na.value = "white"
   )
-#> Error in eval(expr, envir, enclos): object 'temp' not found
 ```
+
+<div class="figure">
+<img src="./faceted-1.png" alt="A faceted SpatRaster" width="100%" />
+<p class="caption">A faceted SpatRaster</p>
+</div>
 
 
 ```r
@@ -201,7 +232,6 @@ ggplot() +
 
 f_volcano <- system.file("extdata/volcano2.tif", package = "tidyterra")
 volcano2 <- rast(f_volcano)
-#> Error: [rast] filename is empty. Provide a valid filename
 
 ggplot() +
   geom_spatraster(data = volcano2) +
@@ -209,10 +239,12 @@ ggplot() +
   scale_fill_whitebox_c() +
   coord_sf(expand = FALSE) +
   labs(fill = "elevation")
-#> Error in `geom_spatraster()`:
-#> ! `tidyterra::geom_spatraster()` only works with <SpatRaster> objects, not <matrix/array>. See
-#>   `?terra::vect()`
 ```
+
+<div class="figure">
[…]
