+++ dev/pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   palettes: palettes.html
   wild: wild.html
   tidyterra: tidyterra.html
-last_built: 2023-07-17T23:02Z
+last_built: 2023-07-23T18:56Z
 urls:
   reference: https://dieghernan.github.io/tidyterra/reference
   article: https://dieghernan.github.io/tidyterra/articles
