+++ vignettes/paper.bib
@@ -1,0 +1,227 @@
+@misc{bahlburg2023,
+	title        = {Intercomparison of growth models for {Antarctic} krill ({Euphausia superba}): towards a generalised understanding},
+	author       = {Dominik Bahlburg and Sally Thorpe and Bettina Meyer and Uta Berger and Eugene Murphy},
+	year         = 2023,
+	journal      = {bioRxiv},
+	publisher    = {Cold Spring Harbor Laboratory},
+	doi          = {10.1101/2023.05.09.540012},
+	url          = {https://www.biorxiv.org/content/early/2023/05/12/2023.05.09.540012},
+	elocation-id = {2023.05.09.540012},
+	abstract     = {Antarctic krill is a key species of the Southern Ocean, impacted by climate change and human exploitation. Understanding how these changes affect the distribution and abundance of krill is crucial for generating projections of change for Southern Ocean ecosystems. Krill growth is an important indicator of habitat suitability and a series of models have been developed and used to examine krill growth potential at different spatial and temporal scales. The available models have been developed using a range of empirical and mechanistic approaches, which provide alternative perspectives and cross-validation of analyses of the key processes influencing krill growth. Here we undertake an intercomparison of a suite of the available models to understand their sensitivities to major driving variables. This illustrates that the results are strongly determined by the model structure and technical characteristics, and the data on which they were developed and validated. Our results emphasize the importance of assessing the constraints and requirements of individual krill growth models to ensure their appropriate application. The study also demonstrates the value of the development of alternative modelling approaches to identify key processes affecting the dynamics of krill. Of critical importance for modelling the growth of krill is appropriately assessing and accounting for differences in estimates of food availability resulting from alternative methods of observation. We suggest that an intercomparison approach is particularly valuable in the development and application of models for the assessment of krill growth potential at circumpolar scales and for future projections. As another result of the intercomparison, the implementations of the models used in this study are now publicly available for future use and analyses.Competing Interest StatementThe authors have declared no competing interest.},
+	eprint       = {https://www.biorxiv.org/content/early/2023/05/12/2023.05.09.540012.full.pdf}
+}
+@book{moraga2023,
+	title        = {Spatial {Statistics} for {Data} {Science}: {Theory} and {Practice} with {R}},
+	shorttitle   = {Spatial {Statistics} for {Data} {Science}},
+	author       = {Moraga, Paula},
+	year         = 2023,
+	publisher    = {Chapman \& Hall/CRC},
+	url          = {https://www.paulamoraga.com/book-spatial/},
+	urldate      = {2023-07-17},
+	note         = {Online available, publication pending},
+	abstract     = {A book for Spatial Statistics for Data Science with R.},
+	language     = {en}
+}
+@manual{R-biomod2,
+	title        = {{biomod2}: Ensemble Platform for Species Distribution Modeling},
+	author       = {Wilfried Thuiller and Damien Georges and Maya Gueguen and Robin Engler and Frank Breiner and Bruno Lafourcade and Remi Patin},
+	year         = 2023,
+	note         = {R package version 4.2-4}
+}
+@misc{R-dplyr,
+	title        = {{dplyr}: A Grammar of Data Manipulation},
[…]
