+++ vignettes/welcome.Rmd
@@ -1,0 +1,408 @@
+---
+title: "Welcome to tidyterra"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Welcome to tidyterra}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+
+## Welcome to {tidyterra}
+
+**tidyterra** is a package that add common methods from the
+[tidyverse](https://www.tidyverse.org/) for SpatRaster and SpatVectors objects
+created with the [{terra}](https://CRAN.R-project.org/package=terra) package. It
+also adds specific `geom_spat*()` functions for plotting these kind of objects
+with [{ggplot2}](https://ggplot2.tidyverse.org/).
+
+### Why {tidyterra}?
+
+Spat\* objects are not like regular data frames. They are a different type of
+objects, implemented via the [S4 object system](http://adv-r.had.co.nz/S4.html),
+and have their own syntax and computation methods, implemented on the {terra}
+package.
+
+By implementing tidyverse methods for these objects, and more specifically
+{dplyr} and {tidyr} methods, a use**R** can now work more easily with Spat\*,
+just like (s)he would do with tabular data.
+
+**Note that** in terms of performance, {terra} is much more optimized for
+working for this kind of objects, so it is **recommended** also to learn a bit
+of {terra} syntax. Each function of {tidyterra} refers (when possible) to the
+corresponding equivalent on {terra}.
+
+## A note for advanced {terra} users
+
+As previously mentioned, {tidyterra} is not optimized in terms of performance.
+Specially when working with `filter()` and `mutate()` methods, it can be slow.
+
+As a rule of thumb, {tidyterra} can handle objects with less than 10.000.000
+slots of information(i.e., `terra::ncell(a_rast) * terra::nlyr(a_rast) < 10e6`).
+
+## Get started with {tidyterra}
+
+Load {tidyterra} with additional libraries of the {tidyverse}:
+
+
+```r
+library(tidyterra)
+library(dplyr)
+library(tidyr)
+```
+
+Currently, the following methods are available:
+
+| tidyverse method             | SpatVector                    | SpatRaster                                                                                     |
+|------------------------------|-------------------------------|------------------------------------------------------------------------------------------------|
+| `tibble::as_tibble()`        | ✔️                            | ✔️                                                                                             |
+| `dplyr::select()`            | ✔️                            | ✔️ Select layers                                                                               |
+| `dplyr::mutate()`            | ✔️                            | ✔️ Create /modify layers                                                                       |
+| `dplyr::transmute()`         | ✔️                            | ✔️                                                                                             |
+| `dplyr::filter()`            | ✔️                            | ✔️ Modify cells values and (additionally) remove outer cells.                                  |
+| `dplyr::slice()`             | ✔️                            | ✔️ Additional methods for slicing by row and column.                                           |
+| `dplyr::pull()`              | ✔️                            | ✔️                                                                                             |
[…]
