+++ vignettes/tidyterra.Rmd
@@ -1,112 +1,138 @@
 ---
-title: "Welcome to tidyterra"
+title: "Using the tidyverse with terra objects: the tidyterra package"
+author:
+  -  Diego Hernangómez^[Independent Researcher, https://orcid.org/0000-0001-8457-4658]
+date: "2023-07-18"
+bibliography: paper.bib
+link-citations: yes
+editor_options:
+  markdown:
+    wrap: 80
+tags:
+  - R
+  - CRAN
+  - spatial
+  - vector
+  - raster
+  - tidyverse
+  - terra
+  - ggplot2
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Welcome to tidyterra}
+  %\VignetteIndexEntry{Using the tidyverse with terra objects: the tidyterra package}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
 
 
-## Welcome to {tidyterra}
+# Summary
 
-**tidyterra** is a package that add common methods from the
-[tidyverse](https://www.tidyverse.org/) for SpatRaster and SpatVectors objects
-created with the [{terra}](https://CRAN.R-project.org/package=terra) package. It
-also adds specific `geom_spat*()` functions for plotting these kind of objects
-with [{ggplot2}](https://ggplot2.tidyverse.org/).
+`tidyterra` is an R [@r-project] package that allows manipulation of spatial
+data objects as provided by the `terra` package [@R-terra], using the syntax of
+several packages included in the tidyverse [@R-tidyverse], such as `dplyr`
+[@R-dplyr], `tidyr` [@R-tidyr], or `tibble` [@R-tibble].
 
-### Why {tidyterra}?
+Furthermore, `tidyterra` extends the functionality of the `ggplot2` package
+[@R-ggplot2] by providing additional geoms and stats [^1], as well as carefully
+chosen scales and color palettes specifically designed for map production.
 
-Spat\* objects are not like regular data frames. They are a different type of
-objects, implemented via the [S4 object system](http://adv-r.had.co.nz/S4.html),
-and have their own syntax and computation methods, implemented on the {terra}
-package.
+[^1]: The term "geoms" refers to geometric objects, and "stats" refers to
+    statistical transformations, following the naming conventions of `ggplot2`
 
-By implementing tidyverse methods for these objects, and more specifically
-{dplyr} and {tidyr} methods, a use**R** can now work more easily with Spat\*,
-just like (s)he would do with tabular data.
+`tidyterra` can manipulate the following classes of `terra` objects:
 
-**Note that** in terms of performance, {terra} is much more optimized for
-working for this kind of objects, so it is **recommended** also to learn a bit
-of {terra} syntax. Each function of {tidyterra} refers (when possible) to the
-corresponding equivalent on {terra}.
+1.  `SpatVector` objects, which represent vector data such as points, lines, or
+    polygon geometries.
 
-## A note for advanced {terra} users
+2.  `SpatRaster` objects, which represent raster data in the form of a grid
+    consisting of equally sized rectangles. Each rectangle can contain one or
+    more values.
 
-As previously mentioned, {tidyterra} is not optimized in terms of performance.
-Specially when working with `filter()` and `mutate()` methods, it can be slow.
+The first stable version of `tidyterra` was included on CRAN on April 24, 2022,
+and has been actively used by other packages (@R-ebvcube, @R-biomod2,
+@R-inlabru, @R-rczechia, @R-sparrpowr) and cited in academic research and
+publications (@bahlburg2023, @moraga2023) ever since.
 
[…]
