+++ pkgdown.yml
@@ -5,7 +5,8 @@ articles:
   faqs: faqs.html
   palettes: palettes.html
   tidyterra: tidyterra.html
-last_built: 2023-07-22T21:56Z
+  welcome: welcome.html
+last_built: 2023-07-31T07:08Z
 urls:
   reference: https://dieghernan.github.io/tidyterra/reference
   article: https://dieghernan.github.io/tidyterra/articles
