+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   palettes: palettes.html
   tidyterra: tidyterra.html
   welcome: welcome.html
-last_built: 2023-07-31T07:08Z
+last_built: 2023-07-31T07:20Z
 urls:
   reference: https://dieghernan.github.io/tidyterra/reference
   article: https://dieghernan.github.io/tidyterra/articles
