+++ vignettes/paper.bib
@@ -140,7 +140,7 @@
 	title        = {Advanced {R}},
 	author       = {Wickham, Hadley},
 	year         = 2019,
-	month        = {05},
+	month        = {5},
 	publisher    = {Chapman and Hall/CRC},
 	doi          = {10.1201/9781351201315},
 	url          = {http://dx.doi.org/10.1201/9781351201315},
@@ -151,7 +151,7 @@
 	title        = {S3},
 	author       = {Wickham, Hadley},
 	year         = 2019,
-	month        = {05},
+	month        = {5},
 	booktitle    = {Advanced {R}},
 	publisher    = {Chapman and Hall/CRC},
 	doi          = {10.1201/9781351201315},
@@ -162,7 +162,7 @@
 }
 @article{ortiz2011organizacion,
 	title        = {La organización territorial de {España} a lo largo de la historia.},
-	author       = {Ruiz Ortíz, Miguel Ángel Ruiz},
+	author       = {Ruiz Ortíz, Miguel Ángel},
 	year         = 2011,
 	journal      = {Revista de Claseshistoria},
 	publisher    = {Clasesdehistoria},
@@ -217,10 +217,10 @@
 	title        = {The Development and Rationale of Cross-blended Hypsometric Tints},
 	author       = {Patterson, Tom and Jenny, Bernhard},
 	year         = 2011,
-	month        = {Jun.},
+	month        = {6},
 	journal      = {Cartographic Perspectives},
 	number       = 69,
-	pages        = {31–46},
+	pages        = {31-46},
 	doi          = {10.14714/CP69.20},
 	url          = {https://cartographicperspectives.org/index.php/journal/article/view/cp69-patterson-jenny},
 	abstractnote = {Hypsometric tints have been a favored mapping technique for over 150 years.By the mid-twentieth century, hypsometric tints based on the work of JohnBartholomew, Jr., Eduard Imhof, and Karl Peucker became the de facto standardfor physical reference maps at small scales. More recently, the role and design ofhypsometric tints have come under scrutiny. One reason for this is the concern thatpeople misread elevation colors as climate or vegetation information. Cross-blendedhypsometric tints, introduced in 2009, are a partial solution to this problem. Theyuse variable lowland colors customized to match the differing natural environmentsof world regions, which merge into one another. In the short time since theirintroduction, cross-blended hypsometric tints have proved to be a popular choiceamong professional mapmakers. Most maps made with cross-blended hypsometrictints also contain shaded relief (terrain represented with modulated light andshadows).}
