+++ vignettes/welcome.Rmd
@@ -7,11 +7,13 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+<!-- welcome.Rmd is generated from welcome.Rmd.orig. Please edit that file -->
+
 
 
 ## Welcome to {tidyterra}
 
-**tidyterra** is a package that add common methods from the
+**tidyterra** is a package that adds common methods from the
 [tidyverse](https://www.tidyverse.org/) for SpatRaster and SpatVectors objects
 created with the [{terra}](https://CRAN.R-project.org/package=terra) package. It
 also adds specific `geom_spat*()` functions for plotting these kind of objects
@@ -54,31 +56,31 @@ library(tidyr)
 
 Currently, the following methods are available:
 
-| tidyverse method             | SpatVector                    | SpatRaster                                                                                     |
-|------------------------------|-------------------------------|------------------------------------------------------------------------------------------------|
-| `tibble::as_tibble()`        | ✔️                            | ✔️                                                                                             |
-| `dplyr::select()`            | ✔️                            | ✔️ Select layers                                                                               |
-| `dplyr::mutate()`            | ✔️                            | ✔️ Create /modify layers                                                                       |
-| `dplyr::transmute()`         | ✔️                            | ✔️                                                                                             |
-| `dplyr::filter()`            | ✔️                            | ✔️ Modify cells values and (additionally) remove outer cells.                                  |
-| `dplyr::slice()`             | ✔️                            | ✔️ Additional methods for slicing by row and column.                                           |
-| `dplyr::pull()`              | ✔️                            | ✔️                                                                                             |
-| `dplyr::rename()`            | ✔️                            | ✔️                                                                                             |
-| `dplyr::relocate()`          | ✔️                            | ✔️                                                                                             |
-| `dplyr::distinct()`          | ✔️                            |                                                                                                |
-| `dplyr::arrange()`           | ✔️                            |                                                                                                |
-| `dplyr::glimpse()`           | ✔️                            | ✔️                                                                                             |
-| `dplyr::inner_join()` family | ✔️                            |                                                                                                |
-| `dplyr::summarise()`         | ✔️                            |                                                                                                |
-| `dplyr::group_by()` family   | ✔️                            |                                                                                                |
-| `dplyr::rowwise()`           | ✔️                            |                                                                                                |
-| `dplyr::count()`, `tally()`  | ✔️                            |                                                                                                |
+| tidyverse method                            | SpatVector                                    | SpatRaster                                                                                     |
+|------------------|------------------|---------------------------------------------|
+| `tibble::as_tibble()`                       | ✔️                                            | ✔️                                                                                             |
+| `dplyr::select()`                           | ✔️                                            | ✔️ Select layers                                                                               |
+| `dplyr::mutate()`                           | ✔️                                            | ✔️ Create /modify layers                                                                       |
[…]
