+++ man/compare_spatrasters.Rd
@@ -35,7 +35,7 @@ On \strong{non-equal crs}, try \code{\link[terra:project]{terra::project()}}. On
 \examples{
 library(terra)
 
-x <- rast(matrix(1:90, ncol = 3), crs = "epsg:3857")
+x <- rast(matrix(1:90, ncol = 3), crs = "EPSG:3857")
 
 # Nothing
 compare_spatrasters(x, x)
