+++ dev/pkgdown.yml
@@ -6,7 +6,8 @@ articles:
   palettes: palettes.html
   wild: wild.html
   tidyterra: tidyterra.html
-last_built: 2023-07-23T18:56Z
+  welcome: welcome.html
+last_built: 2023-07-31T09:17Z
 urls:
   reference: https://dieghernan.github.io/tidyterra/reference
   article: https://dieghernan.github.io/tidyterra/articles
