+++ .github/workflows/lint.yaml
@@ -1,30 +1,68 @@
-# Workflow derived from https://github.com/r-lib/actions/tree/master/examples
-# Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
+# This workflow uses actions that are not certified by GitHub.
+# They are provided by a third-party and are governed by
+# separate terms of service, privacy policy, and support
+# documentation.
+# lintr provides static code analysis for R.
+# It checks for adherence to a given style,
+# identifying syntax errors and possible semantic issues,
+# then reports them to you so you can take action.
+# More details at https://lintr.r-lib.org/
+
+name: lintr
+
 on:
   push:
-    branches: [main, master]
+    branches:
+      - main
+      - master
+  workflow_dispatch:
   pull_request:
-    branches: [main, master]
+    # The branches below must be a subset of the branches above
+    branches:
+      - main
+      - master
+  schedule:
+    - cron: '24 10 * * 3'
 
-name: lint
+permissions:
+  contents: read
 
 jobs:
-  lint:
+  lintr:
+    name: Run lintr scanning
     runs-on: ubuntu-latest
-    env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-    steps:
-      - uses: actions/checkout@v3
+    permissions:
+      contents: read # for checkout to fetch code
+      security-events: write # for github/codeql-action/upload-sarif to upload SARIF results
+      actions: read # only required for a private repository by github/codeql-action/upload-sarif to get the Action run status
 
-      - uses: r-lib/actions/setup-r@v2
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Setup R
+        uses: r-lib/actions/setup-r@v2
         with:
           use-public-rspm: true
 
-      - uses: r-lib/actions/setup-r-dependencies@v2
+      - name: Setup lintr
+        uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::lintr
-          needs: lint
+          extra-packages: |
+            local::.
+            any::lintr
 
-      - name: Lint
+      - name: Lint for comments
         run: lintr::lint_package()
         shell: Rscript {0}
+
+      - name: Run lintr for GitHub
+        run: lintr::sarif_output(lintr::lint_dir("."), "lintr-results.sarif")
+        shell: Rscript {0}
+        continue-on-error: true
+
+      - name: Upload analysis results to GitHub
+        uses: github/codeql-action/upload-sarif@v2
+        with:
+          sarif_file: lintr-results.sarif
+          wait-for-processing: true
