+++ .github/workflows/test-coverage.yaml
@@ -27,14 +27,14 @@ jobs:
           needs: coverage
 
       - name: Test coverage
-        run: | 
+        run: |
           covr::codecov(
             quiet = FALSE,
             clean = FALSE,
             install_path = file.path(Sys.getenv("RUNNER_TEMP"), "package")
           )
         shell: Rscript {0}
-        
+
       - name: Show testthat output
         if: always()
         run: |
