+++ vignettes/articles/INPUBS.bib
@@ -138,3 +138,12 @@
 	doi          = {10.21105/joss.05082},
 	url          = {https://doi.org/10.21105/joss.05082}
 }
+@misc{sepulveda_2023,
+	title        = {Exploring Santiago de Chile: Hillshade and Hypsometric Map with TidyTerra and Ragg},
+	author       = {Ignacio Sepulveda},
+	year         = 2023,
+	month        = aug,
+	journal      = {medium.com},
+	url          = {https://medium.com/@ignaciodsg/exploring-santiago-de-chile-hillshade-and-hypsometric-map-with-tidyterra-and-ragg-6929a6c7143},
+	urldate      = {2023-08-03}
+}
