+++ dev/pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   wild: wild.html
   tidyterra: tidyterra.html
   welcome: welcome.html
-last_built: 2023-07-31T09:17Z
+last_built: 2023-08-03T16:50Z
 urls:
   reference: https://dieghernan.github.io/tidyterra/reference
   article: https://dieghernan.github.io/tidyterra/articles
