+++ vignettes/articles/INPUBS.bib
@@ -139,7 +139,8 @@
 	url          = {https://doi.org/10.21105/joss.05082}
 }
 @misc{sepulveda_2023,
-	title        = {Exploring Santiago de Chile: Hillshade and Hypsometric Map with TidyTerra and Ragg},
+	title        = {Exploring {Santiago}
+	de {Chile}: Hillshade and Hypsometric Map with TidyTerra and Ragg},
 	author       = {Ignacio Sepulveda},
 	year         = 2023,
 	month        = aug,
