### 2023-08-04 #5675 @elbeejay
Hi @dieghernan thanks for the submission. The package on the whole looks good and it is clear that it is in use by the community.

For the paper, we tend to favor shorter papers in JOSS and try to leave longer code examples in the package documentation. To this end, do you think you could shorten the "Examples of use" section in the paper - potentially just show the code and plot for the shortest example. It could be nice to combine the other plots into a single figure and reference it as example outputs that can be generated with the help of the `tidyterra` package, and note that the full code to reproduce those plots is available in the package documentation.

I'm going to go ahead and start searching for reviewers now
