### 2023-08-05 #5675 @dieghernan
Hi @elbeejay

Happy to shorten the paper, but I am afraid it won't be before 14th August, I have limited connectivity until then.

Best
