+++ vignettes/articles/INPUBS.bib
@@ -148,3 +148,15 @@
 	url          = {https://medium.com/@ignaciodsg/exploring-santiago-de-chile-hillshade-and-hypsometric-map-with-tidyterra-and-ragg-6929a6c7143},
 	urldate      = {2023-08-03}
 }
+@article{Leonardi2023.07.24.550358,
+	title        = {{tidysdm}: leveraging the flexibility of {tidymodels} for Species Distribution Modelling in {R}},
+	author       = {Michela Leonardi and Margherita Colucci and Andrea Manica},
+	year         = 2023,
+	journal      = {bioRxiv},
+	publisher    = {Cold Spring Harbor Laboratory},
+	doi          = {10.1101/2023.07.24.550358},
+	url          = {https://www.biorxiv.org/content/early/2023/07/26/2023.07.24.550358},
+	elocation-id = {2023.07.24.550358},
+	eprint       = {https://www.biorxiv.org/content/early/2023/07/26/2023.07.24.550358.full.pdf}
+}
+
