+++ dev/pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   wild: wild.html
   tidyterra: tidyterra.html
   welcome: welcome.html
-last_built: 2023-08-03T17:20Z
+last_built: 2023-08-10T10:43Z
 urls:
   reference: https://dieghernan.github.io/tidyterra/reference
   article: https://dieghernan.github.io/tidyterra/articles
