### 2023-08-10 #5751 @dieghernan
Hi @mikejohnson51, @Yingjie4Science:

Thanks for volunteering on this submission. Please note that after the comment of @elbeejay in https://github.com/openjournals/joss-reviews/issues/5675#issuecomment-1665955080 I modified slightly the Examples section (https://github.com/dieghernan/tidyterra/commit/0820452a2b3ba34307a8c89996c8a44bdfcf48b6) and now the paper wordcount is 907 (therefore between 250-1000 as per the [JOSS Guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).

The corresponding pdf have been generated and linked in https://github.com/openjournals/joss-reviews/issues/5751#issuecomment-1673224096.
