+++ pkgdown.yml
@@ -2,11 +2,11 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
-  faqs: faqs.html
-  palettes: palettes.html
+  xtra_faqs: xtra_faqs.html
+  xtra_palettes: xtra_palettes.html
   tidyterra: tidyterra.html
   welcome: welcome.html
-last_built: 2023-08-01T21:04Z
+last_built: 2023-08-10T14:48Z
 urls:
   reference: https://dieghernan.github.io/tidyterra/reference
   article: https://dieghernan.github.io/tidyterra/articles
