+++ revdep/README.md
@@ -1,66 +1,65 @@
 # Platform
 
-|field    |value                                    |
-|:--------|:----------------------------------------|
-|version  |R version 4.2.2 (2022-10-31 ucrt)        |
-|os       |Windows 10 x64 (build 22621)             |
-|system   |x86_64, mingw32                          |
-|ui       |RStudio                                  |
-|language |(EN)                                     |
-|collate  |Spanish_Spain.utf8                       |
-|ctype    |Spanish_Spain.utf8                       |
-|tz       |Europe/Paris                             |
-|date     |2023-03-17                               |
-|rstudio  |2022.12.0+353 Elsbeth Geranium (desktop) |
-|pandoc   |2.19.2 @ C:\PROGRA~1\Pandoc\pandoc.exe   |
+|field    |value                                                                 |
+|:--------|:---------------------------------------------------------------------|
+|version  |R version 4.3.1 (2023-06-16 ucrt)                                     |
+|os       |Windows Server 2022 x64 (build 20348)                                 |
+|system   |x86_64, mingw32                                                       |
+|ui       |RTerm                                                                 |
+|language |(EN)                                                                  |
+|collate  |English_United States.utf8                                            |
+|ctype    |English_United States.utf8                                            |
+|tz       |UTC                                                                   |
+|date     |2023-08-10                                                            |
+|pandoc   |2.19.2 @ C:\HOSTED~1\windows\pandoc\219~1.2\x64\PANDOC~1.2\pandoc.exe |
 
 # Dependencies
 
 |package      |old    |new    |Δ  |
 |:------------|:------|:------|:--|
-|tidyterra    |0.3.2  |0.4.0  |*  |
-|classInt     |0.4-9  |0.4-9  |   |
-|cli          |3.6.0  |3.6.0  |   |
-|colorspace   |2.1-0  |2.1-0  |   |
-|cpp11        |0.4.3  |0.4.3  |   |
-|data.table   |1.14.8 |1.14.8 |   |
-|DBI          |1.1.3  |1.1.3  |   |
-|dplyr        |1.1.0  |1.1.0  |   |
-|e1071        |1.7-13 |1.7-13 |   |
-|fansi        |1.0.4  |1.0.4  |   |
-|farver       |2.1.1  |2.1.1  |   |
-|generics     |0.1.3  |0.1.3  |   |
-|ggplot2      |3.4.1  |3.4.1  |   |
-|glue         |1.6.2  |1.6.2  |   |
-|gtable       |0.3.1  |0.3.1  |   |
-|isoband      |0.2.7  |0.2.7  |   |
-|labeling     |0.4.2  |0.4.2  |   |
-|lifecycle    |1.0.3  |1.0.3  |   |
-|magrittr     |2.0.3  |2.0.3  |   |
-|munsell      |0.5.0  |0.5.0  |   |
-|pillar       |1.8.1  |1.8.1  |   |
[…]
