### 2023-08-11 #5675 @espinielli
Dear Jay,
I am only seldomnly using terra and a very shallow GIS expert.
I do not think I am up to carry this review.
Sorry for declining your request.

Yours sincerely
Enrico
