+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   xtra_palettes: xtra_palettes.html
   tidyterra: tidyterra.html
   welcome: welcome.html
-last_built: 2023-08-12T21:33Z
+last_built: 2023-08-12T21:46Z
 urls:
   reference: https://dieghernan.github.io/tidyterra/reference
   article: https://dieghernan.github.io/tidyterra/articles
