+++ vignettes/tidyterra.Rmd
@@ -2,6 +2,8 @@
 title: "Using the tidyverse with terra objects: the tidyterra package"
 author: <a href="https://orcid.org/0000-0001-8457-4658">Diego Hernangómez <img alt="ORCID logo" src="https://info.orcid.org/wp-content/uploads/2019/11/orcid_24x24.png" width="16" height="16" /></a>
 date: "2023-07-18"
+description: >-
+  An introduction to the `tidyterra` **R** package.
 bibliography: paper.bib
 link-citations: yes
 editor_options:
@@ -23,12 +25,17 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  warning = FALSE,
+  message = FALSE,
+  out.width = "100%"
+)
+```
 
-
-<!-- tidyterra.Rmd is generated from tidyterra.Rmd.orig. Please edit that file -->
-
-
-# Summary
+## Summary
 
 `tidyterra` is an R [@r-project] package that allows manipulation of spatial
 data objects as provided by the `terra` package [@R-terra], using the syntax of
@@ -54,9 +61,9 @@ chosen scales and color palettes specifically designed for map production.
 The first stable version of `tidyterra` was included on CRAN on April 24, 2022,
 and has been actively used by other packages (@R-ebvcube, @R-biomod2,
 @R-inlabru, @R-rczechia, @R-sparrpowr) and cited in academic research and
-publications (@bahlburg2023, @moraga2023) ever since.
+publications (@bahlburg2023, @moraga2023, @Leonardi2023) ever since.
 
-# Statement of need
+## Statement of need
 
 The [tidyverse](https://www.tidyverse.org/) is a compilation of R packages that
 share an underlying design philosophy, grammar, and data structures. The
@@ -95,7 +102,7 @@ designed for representing spatial phenomena [@whitebox]. Additionally, it
 implements the cross-blended hypsometric tints described by
 @Patterson_Jenny_2011.
 
-# A note on performance
+### A note on performance
 
 The development philosophy of `tidyterra` consists on adapting `terra` objects
 to data frame-like structures by performing different data transformations, that
@@ -108,15 +115,13 @@ handling this type of files.
 Note also that when possible, the help page of each function of `tidyterra`
 references its equivalent in `terra`.
 
-# Examples of use
+## Example of use
 
 `tidyterra` is available on
 [CRAN](https://CRAN.R-project.org/package=tidyterra), so it can be easily
 installed using the following commands in R:
 
-
-
-```r
+```{r eval=FALSE}
 install.packages("tidyterra")
 ```
 
@@ -124,9 +129,7 @@ The following example demonstrates how to manipulate a `SpatRaster` object using
 the `dplyr` syntax. Additionally, it illustrates how to seamlessly plot a
 `SpatRaster` object with `ggplot2` using the `geom_spatraster()` function:
 
-
-
-```r
+```{r eval=FALSE}
 library(tidyterra)
 library(tidyverse) # Load all the packages of tidyverse at once
 library(scales) # Additional library for labels
@@ -135,17 +138,6 @@ library(scales) # Additional library for labels
 rastertemp <- terra::rast(system.file("extdata/cyl_temp.tif",
   package = "tidyterra"
 ))
-
-rastertemp
-#> class       : SpatRaster
-#> dimensions  : 87, 118, 3  (nrow, ncol, nlyr)
-#> resolution  : 3881.255, 3881.255  (x, y)
[…]
