+++ .github/workflows/pkgdown-gh-pages-clean.yaml
@@ -32,3 +32,4 @@ jobs:
           git config --local user.name "github-actions[bot]"
           git config --local user.email "41898282+github-actions[bot]@users.noreply.github.com"
           Rscript -e 'pkgdown::deploy_to_branch(new_process = FALSE, clean = TRUE)'
+
