+++ .github/workflows/cran-status-check.yaml
@@ -8,16 +8,17 @@ on:
 jobs:
   check:
     runs-on: ubuntu-latest
-    permissions: write-all 
+    permissions: write-all
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
     steps:
       - name: Checkout
         uses: actions/checkout@v3
-        
+
       - name: Check
         uses: dieghernan/cran-status-check@v1
         with:
           fail-on-error: "true"
-          create-issue: "false" 
+          create-issue: "false"
+
