+++ pkgdown/_pkgdown.yml
@@ -117,3 +117,4 @@ reference:
 - title: "About the package"
   contents:
     - tidyterra-package
+
