+++ R/scales_hypso.R
@@ -14,7 +14,8 @@
 #'   [grDevices::terrain.colors()] for details.
 #'
 #' An additional set of scales is provided. These scales can act as
-#' [hypsometric (or bathymetric) tints](https://en.wikipedia.org/wiki/Hypsometric_tints).
+#' [hypsometric (or bathymetric)
+#' tints](https://en.wikipedia.org/wiki/Hypsometric_tints).
 #'
 #' - `scale_*_hypso_tint_d()`: For discrete values.
 #' - `scale_*_hypso_tint_c()`: For continuous values.
