+++ tests/testthat/test-scales_fill_whitebox.R
@@ -36,9 +36,9 @@ test_that("Discrete scale", {
   mod_alpha_rev <- ggplot2::layer_data(p4)$fill
 
 
-  expect_true(all(rev(alpha(mod,
-    alpha = 0.7
-  )) == mod_alpha_rev))
+  expect_true(
+    all(rev(alpha(mod, alpha = 0.7)) == mod_alpha_rev)
+  )
 
   # Change pal
   # Create ggplot for each pal, extract colors and check
@@ -103,9 +103,9 @@ test_that("Continous scale", {
   mod_alpha_rev <- ggplot2::layer_data(p4)$fill
 
 
-  expect_true(all(rev(alpha(mod,
-    alpha = 0.7
-  )) == mod_alpha_rev))
+  expect_true(
+    all(rev(alpha(mod, alpha = 0.7)) == mod_alpha_rev)
+  )
 
   # Change pal
   # Create ggplot for each pal, extract colors and check
