+++ tests/testthat/test-scales_fill_wiki.R
@@ -36,9 +36,11 @@ test_that("Discrete scale", {
   mod_alpha_rev <- ggplot2::layer_data(p4)$fill
 
 
-  expect_true(all(rev(alpha(mod,
-    alpha = 0.7
-  )) == mod_alpha_rev))
+  expect_true(
+    all(
+      rev(alpha(mod, alpha = 0.7)) == mod_alpha_rev
+    )
+  )
 })
 
 test_that("Continous scale", {
