+++ tests/testthat/test-scales_col_wiki.R
@@ -43,9 +43,9 @@ test_that("Discrete scale", {
   mod_alpha_rev <- ggplot2::layer_data(p4)$colour
 
 
-  expect_true(all(rev(alpha(mod,
-    alpha = 0.7
-  )) == mod_alpha_rev))
+  expect_true(
+    all(rev(alpha(mod, alpha = 0.7)) == mod_alpha_rev)
+  )
 })
 
 test_that("Continous scale", {

[…]
