+++ .github/workflows/check-terra-devel.yaml
@@ -43,13 +43,15 @@ jobs:
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
           cache-version: terra-devel
-          extra-packages: any::rcmdcheck
+          extra-packages: |
+            any::rcmdcheck
+            rspatial/terra
           needs: check
 
       - name: Install terra from runiverse
         run: |
           # Update packages  repository from rspatial runiverse
-          update.packages(repos = 'https://rspatial.r-universe.dev')
+          # update.packages(repos = 'https://rspatial.r-universe.dev')
 
           # Display version
           cat(as.character(packageVersion("terra")))
