+++ .github/workflows/wipe-cache.yaml
@@ -1,0 +1,11 @@
+name: Clear GHA caches manually
+on:
+  workflow_dispatch:
+
+jobs:
+  cache-clear:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: easimon/wipe-cache@main
+
