+++ vignettes/articles/INPUBS.bib
@@ -161,4 +161,13 @@
 	elocation-id = {2023.07.24.550358},
 	eprint       = {https://www.biorxiv.org/content/early/2023/07/26/2023.07.24.550358.full.pdf}
 }
-
+@book{desaix_2023,
+	title        = {Migratory Networks in {R}},
+	author       = {DeSaix, Matt},
+	year         = 2023,
+	month        = aug,
+	publisher    = {Online},
+	url          = {https://mgdesaix.github.io/connectivity-book/},
+	urldate      = {2023-09-01},
+	language     = {en}
+}
