+++ DESCRIPTION
@@ -44,7 +44,7 @@ VignetteBuilder:
     knitr
 Config/Needs/coverage: covr
 Config/Needs/website: geodata, metR, dieghernan/gitdevr, ragg, styler,
-    Nowosad/spDataLarge
+    Nowosad/spDataLarge, ggspatial
 Config/testthat/edition: 3
 Config/testthat/parallel: true
 Encoding: UTF-8
