### 2023-10-08 dieghernan/tidyterra#116 @Yingjie4Science
JOSS review
This new post is created for JOSS review https://github.com/openjournals/joss-reviews/issues/5751. In general, this is a well-developed R package and the key functions have been well-documented. In addition to the checklist, I have a few minor comments on the draft, and I hope this could be useful for improving the readability of your manuscript, especially to general audience.

- [ ] Summary: line 11 "by providing additional geoms and stats", could you specify the names of these additions? Similarly, you mentioned this package "has been actively used by other packages", but did not mention the package names, though you added the citations.
- [ ] Summary: perhaps you can add one short sentence on the broad application or implication of this new package. This is not super clear to me and potentially other readers who first hear about it.
- [ ] 34-36 " `terra` offers a wide set of functions with their own syntax for
transforming and visualizing" - could you specify what are the drawbacks of using "functions with their own syntax", and how the integration can address that? I see the following paragraph was trying to answer this question but still, it is not very clear to me.
