### 2023-10-13 dieghernan/tidyterra#117 @mikejohnson51
JOSS Review: Johnson
@dieghernan what a cool package you've made. I use it regularly and am happy to be a part of this review process. I think the paper and page website stand on their own and are good as is. Below are few points that I think would help better convey the utility you are bringing to the R spatial ecosystem. All are optional in my mind to be considered for publication in JOSS.

# Statement of need:

- I would add a little more about `rasterVis` and `stars` plotting and how you provide similar key concepts in a widely used plotting package (`ggplot`) for the more modern spatial data structure (`terra`) and data manipulation methods (`dplyr`). You are in many ways providing the core of what `sf` gave vector data to the `terra` world! The text dances around this but I think what you have done allows you to be more direct.
- The popularity and utility of `rasterVis` as an supplementary package to `raster` is somewhat analogous to this paradigm and speaks to the need for more advanced plotting methods, and the value of an addition in the `ggplot2` ecosystem
- Your current text makes it seem like the primary addition (over rasterVis and `ggspatial`) is `SpatVect` integration (see example of use point 3) but equally important (I think) is that (most) R users do not have to learn a new plotting syntax to engage with raster data. That is super powerful!  When I picked up yours, it was instantly familiar, In not slight to those other packages, I regularly had to bring up the help pages to get what I wanted. I dont suspect I am alone in this and the familiarity is what makes `tidyterra` cool.

# Performance note

  - Out of scope for this, but in the future having some auto resampling like `stars` would be slick for plotting. If the performace issues are only related to data manipulation then Id state this explicilty.

# Example of use

- Since this is an actively maintained package, I would give the `remotes::install_github()` option as well
- Personally one of the things I love most about your package is the palettes you added. I would prefer to see those mentioned in the text. A broader point is that the supplementary website is wonderful, I would like to see the Articles highlighted in the paper text with hyperlinks if possible.
- Adding an additional brief example using a `SpatVect `object and more traditional `dplyr` (mutate/select/filter) example would be really powerful to show how you’ve been able to add much of what makes `sf` so unique, to the `terra` paradigm. As it stands, the `rename` example is ok, but weak compared to the power your package offers.
