### 2023-10-13 #5751 @mikejohnson51
@elbeejay @dieghernan I have completed my review and added minor comments as a new [issue](https://github.com/dieghernan/tidyterra/issues/117)
