### 2023-10-23 #5751 @elbeejay
Hi @dieghernan yes that's no problem. The paper can be extended a bit as needed, the guidance for paper length is to keep them from being too long, but is not a hard rule.
