### 2023-10-28 dieghernan/tidyterra#116 @dieghernan
Thanks for your review @Yingjie4Science

I think I tackled all of them, see changes on https://github.com/dieghernan/tidyterra/commit/77a268d3e729d922690f51472f7b31b7be94f33c

Regards
