### 2023-10-29 dieghernan/tidyterra#116 @Yingjie4Science
@dieghernan Thank you for addressing these comments. It looks great!
