### 2023-10-30 dieghernan/tidyterra#117 @dieghernan
Hi @mikejohnson51:

First of all, thanks for this great feedback. I did some modifications in line with your comments, however I assumed also your statement considering all of them optional since currently I am not able to dedicate much time due to personal reasons. See below a short summary of the changes (see also https://github.com/dieghernan/tidyterra/compare/77a268d..144c6ab)
- I tried to emphasize other features of `tidyterra` vs `rasterVis/ggspatial`, as the capacity of create faceted maps (that I think is really useful in the context of SpatRasters), contours, auto-conversion of spatial layers to the same CRS and by last the support for SpatVector objects.
- I made clear that the performance may be low on the manipulation as you already guessed. The solution you proposed for plotting (re-sampling) is efectively implemented since the initial version of the package in line with `terra::plot()`.
- Agree on your view of the example, I included an additional one with a mutate/select worflow, the overlay of a SpatVector
and a contour plot of the result, trying to present a more complex map.
- I added a final section highlighting the documentation and the vignettes, including a explicit reference to the article that showcases all the palettes.

Best, Diego
