### 2023-10-30 #5751 @dieghernan
Hi @elbeejay

I already made changes in the paper in line with https://github.com/dieghernan/tidyterra/issues/116 and https://github.com/dieghernan/tidyterra/issues/117 (thanks @Yingjie4Science  and @mikejohnson51  for your time and useful feedback).
