+++ vignettes/articles/INPUBS.bib
@@ -171,3 +171,16 @@
 	urldate      = {2023-09-01},
 	language     = {en}
 }
+@article{meister2023,
+	title        = {Blue mussels in western Norway have vanished where in reach of crawling predators},
+	author       = {Nadja Meister and Tom J. Langbehn and  Øystein Varpe and Christian Jørgensen},
+	year         = 2023,
+	month        = 10,
+	journal      = {Marine Ecology Progress Series},
+	volume       = 721,
+	pages        = {85--101},
+	doi          = {10.3354/meps14416},
+	url          = {http://dx.doi.org/10.3354/meps14416},
+	date         = {2023-10-19},
+	langid       = {en}
+}
