### 2023-10-31 #5751 @elbeejay
Great, looks like you can close https://github.com/dieghernan/tidyterra/issues/116, thanks @Yingjie4Science for following up there and here on your reviewer checklist.

@mikejohnson51 can you check out the revisions @dieghernan made in https://github.com/dieghernan/tidyterra/issues/117 in response to your comments? Thanks!
