### 2023-11-01 dieghernan/tidyterra#117 @mikejohnson51
I really like the new example. Thanks for making the time! I think this looks great for JOSS.
