### 2023-11-06 #5751 @elbeejay
Great, thanks for your comment in the issue @mikejohnson51, can you check off your remaining checklist items above?

@dieghernan it looks like you can close https://github.com/dieghernan/tidyterra/issues/117
