### 2023-11-07 dieghernan/tidyterra#117 @dieghernan
Closed, thanks!
