### 2023-11-07 #5751 @dieghernan
Hi @elbeejay:

I completed the checklist accordingly, the released version is **v0.4.1** and the corresponding DOI is https://doi.org/10.5281/zenodo.6572471.

Let me know if everything is correct, many thanks.
