+++ tests/testthat/_snaps/stat_spatraster.md
@@ -2,13 +2,15 @@
 
     Code
       ggplot() + stat_spatraster(data = v)
-    Error <rlang_error>
-      `tidyterra::stat_spatraster()` only works with <SpatRaster> objects, not <SpatVector>. See `?terra::vect()`
+    Condition
+      Error in `stat_spatraster()`:
+      ! `tidyterra::stat_spatraster()` only works with <SpatRaster> objects, not <SpatVector>. See `?terra::vect()`
 
 ---
 
     Code
       ggplot() + stat_spatraster(data = 1:3)
-    Error <rlang_error>
-      `tidyterra::stat_spatraster()` only works with <SpatRaster> objects, not <integer>. See `?terra::vect()`
+    Condition
+      Error in `stat_spatraster()`:
+      ! `tidyterra::stat_spatraster()` only works with <SpatRaster> objects, not <integer>. See `?terra::vect()`
 
