+++ tests/testthat/_snaps/pull_crs.md
@@ -2,7 +2,7 @@
 
     Code
       res <- pull_crs(list(1))
-    Message <cliMessage>
+    Message
       ! On `tidyterra::pull_crs()`
       No wkt equivalent found. Returning NA
 
