+++ tests/testthat/_snaps/autoplot-Spat.md
@@ -2,7 +2,7 @@
 
     Code
       b <- ggplot2::ggplot_build(forced)
-    Message <cliMessage>
+    Message
       ! `tidyterra::geom_spatraster()`: Plotting 4 overlapping layers: tavg_04, tavg_05, tavg_06, and other. Either:
         Use `facet_wrap(~lyr)` for faceting or
         Use `aes(fill = <name_of_layer>)` for displaying single layers

[…]
