### 2023-11-08 1be75cf
Make get_coltab_pal robust when layers does not have a unique name
