+++ README.Rmd
@@ -83,7 +83,7 @@ example, `as_tibble()` would return a tibble).
 Current methods and functions provided by {tidyterra} are:
 
 | tidyverse method                            | SpatVector                                    | SpatRaster                                                                                     |
-|---------------------------------------------|-----------------------------------------------|------------------------------------------------------------------------------------------------|
+|--------------------|--------------------|----------------------------------------|
 | `tibble::as_tibble()`                       | ✔️                                            | ✔️                                                                                             |
 | `dplyr::select()`                           | ✔️                                            | ✔️ Select layers                                                                               |
 | `dplyr::mutate()`                           | ✔️                                            | ✔️ Create /modify layers                                                                       |
