### 2023-11-09 #5751 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
